completed exam topic create, delete, update

+++ app/Http/Controllers/ExamCategoryController.php
@@ -3,14 +3,17 @@
 namespace App\Http\Controllers;
 
 use App\Models\ExamCategory;
+use Illuminate\Contracts\Foundation\Application;
+use Illuminate\Contracts\View\Factory;
+use Illuminate\Contracts\View\View;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
-use Illuminate\Validation\Validator;
 
 class ExamCategoryController extends Controller
 {
     /**
-     * Load exam model index page to view category list and create new category
-     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
+     * Load model exam index page to view category list and create new category
+     * @return Application|Factory|View
      */
     public function index()
     {
@@ -21,7 +24,7 @@ class ExamCategoryController extends Controller
     /**
      * Store new category data
      * @param Request $request
-     * @return \Illuminate\Http\RedirectResponse
+     * @return RedirectResponse
      */
     public function store(Request $request)
     {
@@ -33,9 +36,28 @@ class ExamCategoryController extends Controller
         return redirect()->back()->with(['status' => 'Category Created Successfully']);
     }
 
+    /**
+     * Delete specific exam category
+     * @param $id
+     * @return RedirectResponse
+     */
     public function destroy($id)
     {
+        if($this->hasTopics($id)) {
+            return redirect()->back()->with(['warning' => 'Can not delete Category! This Category has its associative topics']);
+        }
         ExamCategory::query()->find($id)->delete();
         return redirect()->back()->with(['status' => 'Category Deleted Successfully']);
+    }
+
+    /**
+     * Check if the category has its associative topics
+     * If category has topics, it can not be deleted
+     * @param $categoryId
+     * @return bool
+     */
+    private function hasTopics($categoryId)
+    {
+        return ExamCategory::query()->where('id',$categoryId)->has('examTopics')->exists();
     }
 }
